requirements.txt
langchain>=1.0.0
langgraph
click
rich
psutil
pandas
openpyxl
python-dotenv
PyPDF2
requests
beautifulsoup4
duckduckgo-search
langchain-ollama
pywhatkit
